pyproject.toml
[project]
name = "stream-data-producer"
version = "0.1.0"
description = "A configurable single-stream data generator for testing and simulation"
authors = [
    {name = "Shimeng Zhao", email = "shimengzhao@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "confluent-kafka>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "click>=8.0.0",
    "psutil>=5.9.0",
    "requests>=2.28.0",
    "httpx>=0.23.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[project.scripts]
stream-data-producer = "stream_data_producer.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/stream-data-producer"
Repository = "https://github.com/yourusername/stream-data-producer.git"
Documentation = "https://github.com/yourusername/stream-data-producer#readme"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
